% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.lst.R
\name{extract.lst}
\alias{extract.lst}
\title{extract.lst}
\usage{
extract.lst(ifile, ofile, delete.original = TRUE)
}
\arguments{
\item{ifile}{Path to hdf file.}

\item{ofile}{Base filename of the output. The name will be \emph{ofile} + '_day-lst.tif' and \emph{ofile} + '_night-lst.tif'}

\item{delete.original}{Logical argument specifying if the hdf should be deleted when processed.}
}
\value{
One or multiple raster objects.
}
\description{
Interface to download and process tile-wise Land Surface Temperature (LST) data.
}
\details{
{Downloads and pre-processes 
\link[https://lpdaac.usgs.gov/dataset_discovery/modis/modis_products_table/mod11a2_v006]{MOD11A2} and 
\link[https://lpdaac.usgs.gov/dataset_discovery/modis/modis_products_table/myd11a2_v006]{MYD11A2} data 
for user specified \emph{tiles}. The data is downloaded from the 
\link[https://ladsweb.modaps.eosdis.nasa.gov/]{LAADS DAAC server}. for each tile, the function downloads 
the hdf files for closest in time to the elements in \emph{dates} and, for each hdf file, extract day and night 
LST, applies quality information to each band and stores them as separate files names as "Date" + "tile" + 
"collection" + "variable" + ".tif".}
}
