% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.lst.gaps.R
\name{fill.lst.gaps}
\alias{fill.lst.gaps}
\title{fill.lst.gaps}
\usage{
fill.lst.gaps(x, y, target.year, ofile)
}
\arguments{
\item{x}{\emph{character} vector with the paths of \emph{RasterLayers}.}

\item{y}{an Object of class \emph{Date} with the same length as \emph{x}.}

\item{target.year}{Numeric element specifying the target year (additional dates are only used for interpalation)}

\item{ofile}{Path to output file.}
}
\value{
A \emph{character} vector and hdf files.
}
\description{
Fill temporal gaps in LST \emph{RasterLayer} objects.
}
\details{
The function uses an algorithm similar to the one offered 
by \code{\link[rsMove]{imgInt}}. In essense, the function evaluates 
the time series of each pixel and, for each date with a NA value, it 
searches for the closest, non-NA values in time - contrained by a 
temporal buffer - and uses them to fill data gaps linearly.
}
