% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full.process.R
\name{full.process}
\alias{full.process}
\title{full.process}
\usage{
full.process(tile, dates, data.path1, data.path2)
}
\arguments{
\item{tile}{\emph{character} vector specifying the target MODIS tile (e.g. "h01v01")}

\item{dates}{a vector of class \emph{Date} containing the target download dates.}

\item{data.path1}{Output data path for downloaded data.where tile-wise data will be stored.}

\item{data.path2}{Output data path for downloaded data.where Mosaics will be stored.}
}
\value{
A \emph{character} vector and hdf files.
}
\description{
Full processing of LST including data download, masking, monthly averaging and compositing
}
