% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.mean.lst.R
\name{monthly.mean.lst}
\alias{monthly.mean.lst}
\title{month.mean.lst}
\usage{
monthly.mean.lst(x, y)
}
\arguments{
\item{x}{A \emph{RasterStack}.}

\item{y}{an Object of class \emph{date} with the same length as \emph{x}.}
}
\value{
A \emph{RasterStack}.
}
\description{
Derives monthly-mean LST based on a \emph{RasterStack}.
}
\details{
Builds a \emph{RasterStack} of monthly-mean lST based on e.g. 8-day images.
}
