% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiCore.lst.R
\name{multiCore.lst}
\alias{multiCore.lst}
\title{multiCore.lst}
\usage{
multiCore.lst(tiles, dates, year, data.path1, data.path2)
}
\arguments{
\item{tiles}{\emph{character} vector specifying the target MODIS tiles (e.g. "h01v01")}

\item{dates}{a vector of class \emph{Date} containing the target download dates.}

\item{year}{Numeric element specifying the target year (additional dates are only used for interpalation)}

\item{data.path1}{Output data path for downloaded data.where tile-wise data will be stored.}

\item{data.path2}{Output data path for downloaded data.where Mosaics will be stored.}
}
\value{
One \emph{RasterStack} object for each element in \emph{tiles}.
}
\description{
Multi-core implementation of the \code{\link{pro.lst}} function.
}
