% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro.lst.R
\name{pro.lst}
\alias{pro.lst}
\title{pro.lst}
\usage{
pro.lst(tile, date, data.path)
}
\arguments{
\item{tile}{\emph{character} vector specifying the target MODIS tile (e.g. "h01v01")}

\item{date}{a vector of class \emph{Date} containing the target download dates.}

\item{data.path}{Output data path for downloaded data.}
}
\value{
One or multiple raster objects.
}
\description{
Interface to download and process tile-wise Land Surface Temperature (LST) data.
}
\details{
{Downloads and pre-processes 
\link[https://lpdaac.usgs.gov/dataset_discovery/modis/modis_products_table/mod11a2_v006]{MOD11A2} and 
\link[https://lpdaac.usgs.gov/dataset_discovery/modis/modis_products_table/myd11a2_v006]{MYD11A2} data 
download for a given \emph{tile}. The function downloads the hdf file closest in time to \emph{date} 
data is downloaded from the \link[https://ladsweb.modaps.eosdis.nasa.gov/]{LAADS DAAC server}. Then, it 
extract day and night LST masked with Quanlity Control (QC) data and, when, TERRA and AQUA images are 
available, merges both in a single file for each variable (i.e. "day" and "night"). The output files are 
named as "Date" + "tile" + "product"/"combined" + "variable" + ".tif".}
}
