% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.dates.R
\name{update.dates}
\alias{update.dates}
\title{update.dates}
\usage{
\method{update}{dates}(x)
}
\arguments{
\item{x}{a vector of class \emph{Date} containing the target download dates.}
}
\value{
a \emph{data.frame}.
}
\description{
Finds closest, possible dates to download LST data for.
}
\details{
{Finds closest, possible dates to download 8-day LST data based on a set 
of reference dates. Additionally, the function reports on the year, month and Day 
of Acquisition (DoA).}
}
